+++ src/backends/fs/index.js
@@ -1,5 +1,4 @@
-import winston from 'winston';
-
+import { logger } from '../../config/logger';
 import config from '../../config/config';
 
 // Import required fs interfaces
@@ -7,12 +6,12 @@ const FS_INTERFACES = {};
 const _interfaces = config.FS_INTERFACES.split(',').map(v => v.trim());
 for (const _interface of _interfaces) {
   FS_INTERFACES[_interface] = require(`./${_interface}`);  // eslint-disable-line global-require
-  winston.debug(`Setup '${_interface}' interface`);
+  logger.debug(`Setup '${_interface}' interface`);
 
   // Ensure non-ssl protocols will be handled by https interface
   if (_interface === 'https') {
     FS_INTERFACES.http = FS_INTERFACES.https;
-    winston.debug("Setup 'http' interface");
+    logger.debug("Setup 'http' interface");
   }
 }
 export default FS_INTERFACES;