+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Release History
 
+* 1.1.0 - Support Opbeat logger. Nov 8th, 2017.
+* 1.0.3 - Use NYC for coverage, add Docker tooling. Nov 3rd, 2017.
 * 1.0.2 - Support default bundle name. Aug 28th, 2017.
 * 1.0.1 - Bugfix: Trim S3 keys of preceding/trailing slashes. Aug 18th, 2017.
 * 1.0.0 - Initial release. Aug 17th, 2017.